Safari-611.1.10/LayoutTests/editing/pasteboard/data-transfer-set-data-ignore-copied-malformed-url-in-null-origin-expected.txt

This tests calling setData with a malformed URL in a null origin document. The malformed value should not be readable in another document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS urlReadInSameDocument is "hello, world"
PASS JSON.stringify(typesInSameDocument) is "[\"text/uri-list\"]"
PASS JSON.stringify(itemsInSameDocument) is "[{\"kind\":\"string\",\"type\":\"text/uri-list\"}]"
PASS event.clipboardData.getData("url") is ""
PASS JSON.stringify(event.clipboardData.types) is "[]"
PASS JSON.stringify(Array.from(event.clipboardData.items).map((item) => ({kind: item.kind, type: item.type}))) is "[]"
PASS successfullyParsed is true

TEST COMPLETE
